Application
The application is made through the Django extension in Python. It is a web application that is run on the localhost server. The app uses the inbuilt admin database in Django to store categories and user information which are segregated into three categories: PATIENT, ADMIN and DOCTOR to signify the three divisions at the user's hospital. The app uses simple and clear UI features to aid the admin to perform various tasks, looking at patient records according to calendar dates and searching at the relevant details in the dashboard.